+++ test/support/extensions/email_confirmation/user_identities_mock.ex
@@ -7,6 +7,6 @@ defmodule PowAssent.Test.EmailConfirmation.UserIdentitiesMock do
 
   def get_user_by_provider_uid("test_provider", "new_user-missing_email_confirmation"), do: %{@user | email_confirmed_at: nil}
 
-  def create_user("test_provider", "new_user", _params, %{"email" => "foo@example.com"}), do: {:ok, %{@user | email: "foo@example.com"}}
-  def create_user("test_provider", "new_user", %{"email" => "foo@example.com"}, _user_params), do: {:ok, %{@user | email: "foo@example.com", email_confirmed_at: DateTime.utc_now()}}
+  def create_user(%{"provider" => "test_provider", "uid" => "new_user"}, _params, %{"email" => "foo@example.com"}), do: {:ok, %{@user | email: "foo@example.com"}}
+  def create_user(%{"provider" => "test_provider", "uid" => "new_user"}, %{"email" => "foo@example.com"}, _user_params), do: {:ok, %{@user | email: "foo@example.com", email_confirmed_at: DateTime.utc_now()}}
 end