+++ CHANGELOG.md
@@ -1,9 +1,23 @@
 # Changelog
 
-## v0.2.5 (TBA)
+## v0.3.0 (TBA)
 
 * Added `PowAssent.Phoenix.ViewHelpers.authorization_link/2` and  `PowAssent.Phoenix.ViewHelpers.deauthorization_link/2`
-* Deprecated `PowAssent.Phoenix.ViewHelpers.provider_link/3`
+* Removed `PowAssent.Phoenix.ViewHelpers.provider_link/3`
+* Rewritten plug methods and controller handling so they now pass through additional params such as access token. This makes it possible to e.g. capture access tokens. Now there is a clear distinction between user identity params and user params, and most methods now accepts or returns two separate params. Following methods updated:
+  * `MyApp.UserIdentities.create/3` changed to `MyApp.UserIdentities.create/2`
+  * `MyApp.UserIdentities.pow_assent_create/3` changed to `MyApp.UserIdentities.pow_assent_create/2`
+  * `PowAssent.Ecto.UserIdentities.Context.create/4` changed to `PowAssent.Ecto.UserIdentities.Context.create/3`
+  * `MyApp.UserIdentities.create_user/4` changed to `MyApp.UserIdentities.create_user/3`
+  * `MyApp.UserIdentities.pow_assent_create_user/4` changed to `MyApp.UserIdentities.pow_assent_create_user/3`
+  * `PowAssent.Ecto.UserIdentities.Context.create_user/5` changed to `PowAssent.Ecto.UserIdentities.Context.create_user/4`
+  * `PowAssent.Operations.create/4` changed to `PowAssent.Operations.create/3`
+  * `PowAssent.Operations.create_user/5` changed to `PowAssent.Operations.create_user/4`
+  * `PowAssent.Plug.callback/4` now returns a tuple with `{:ok, user_identity_params, user_params, conn}`
+  * `PowAssent.Plug.authenticate/3` changed to `PowAssent.Plug.authenticate/2`
+  * `PowAssent.Plug.create_identity/3` changed to `PowAssent.Plug.create_identity/2`
+  * `PowAssent.Plug.create_user/4` now accepts `user_identity_params` instead of `provider` as second argument
+  * `PowAssent.Plug.create_user/4` now expects `user_identity_params` rather than `provider` as second argument
 
 ## v0.2.4 (2019-04-25)